# Use rb_type() to check primitive class type