--- a/ext/oj/object.c
+++ b/ext/oj/object.c
@@ -536,7 +536,7 @@ WHICH_TYPE:
         }
         break;
     case T_HASH:
-        if (rb_cHash != rb_obj_class(parent->val)) {
+        if (T_HASH != rb_type(parent->val)) {
             if (4 == klen && 's' == *key && 'e' == key[1] && 'l' == key[2] && 'f' == key[3]) {
                 rb_funcall(parent->val, oj_replace_id, 1, value);
             } else {