It seems slightly faster to use `rb_type()` to check for primitive class types than `rb_obj_class()​`.

−               | before   | after    | result
--               | --       | --       | --
Oj.dump          | 439.927k | 442.567k | 1.006x

### Environment
- MacBook Pro (M1 Max, 2021)
- macOS 12.0
- Apple M1 Max
- Ruby 3.0.2

### Before
```
Warming up --------------------------------------
Oj.load(symbol_keys)    44.036k i/100ms
Calculating -------------------------------------
Oj.load(symbol_keys)    439.927k (± 0.3%) i/s -      4.404M in  10.009941s
```

### After
```
Warming up --------------------------------------
Oj.load(symbol_keys)    44.146k i/100ms
Calculating -------------------------------------
Oj.load(symbol_keys)    442.567k (± 0.3%) i/s -      4.459M in  10.074843s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json =<<-EOF
{
  "$id": "https://example.com/person.schema.json",
  "$schema": "https://json-schema.org/draft/2020-12/schema",
  "title": "Person",
  "type": "object",
  "properties": {
    "firstName": {
      "type": "string",
      "description": "The person's first name."
    },
    "lastName": {
      "type": "string",
      "description": "The person's last name."
    },
    "age": {
      "description": "Age in years which must be equal to or greater than zero.",
      "type": "integer",
      "minimum": 0
    }
  }
}
EOF

Benchmark.ips do |x|
  x.warmup = 10
  x.time = 10

  x.report('Oj.load(symbol_keys)') { Oj.load(json, symbol_keys:true) }
end
```