@ohler55 2021-11-05
It looks like the changes fail due to the difference between an object being a Hash type versus being of the Hash class.